+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/lib/src/types/types.dart
@@ -34,3 +34,4 @@ export 'utils/marker.dart';
 export 'utils/polygon.dart';
 export 'utils/polyline.dart';
 export 'utils/tile_overlay.dart';
+export 'web_gesture_handling.dart';