+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.3.0
 
+* Adds options for gesture handling and tilt controls on web.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.2.3